--- src/app/components/landing-page/landing-page.module.ts
@@ -1,30 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { LandingPageComponent } from './landing-page.component';
-import {TranslateModule, TranslateLoader} from '@ngx-translate/core';
-import {HttpClient} from '@angular/common/http';
-import {TranslateHttpLoader} from '@ngx-translate/http-loader';
-import { ServicesComponent } from 'src/app/services/services.component';
-import { ServicesModule } from 'src/app/services/services.module';
-
-export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
-}
-
-@NgModule({
-  declarations: [
-    LandingPageComponent,
-    ServicesComponent
-  ],
-  imports: [
-    CommonModule,
-    TranslateModule.forRoot({
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
-      }
-    })
-  ]
-})
-export class LandingPageModule { }
