+++ src/app/mainModules/about-us/about-us.component.ts
