+++ src/app/mainModules/landing-page/landing-page.component.spec.ts
