+++ src/app/mainModules/landing-page/landing-page.component.ts
