+++ src/app/mainModules/landing-page/landing-page.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SigninComponent } from './signin.component';
+import { LandingPageComponent } from './landing-page.component';
 import {TranslateModule, TranslateLoader} from '@ngx-translate/core';
 import {HttpClient} from '@angular/common/http';
 import {TranslateHttpLoader} from '@ngx-translate/http-loader';
-import { ReactiveFormsModule } from '@angular/forms';
+import { ServicesModule } from '../services/services.module';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -12,11 +12,11 @@ export function HttpLoaderFactory(http: HttpClient) {
 
 @NgModule({
   declarations: [
-    SigninComponent
+    LandingPageComponent
   ],
   imports: [
     CommonModule,
-    ReactiveFormsModule,
+    ServicesModule,
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
@@ -26,4 +26,4 @@ export function HttpLoaderFactory(http: HttpClient) {
     })
   ]
 })
-export class SigninModule { }
+export class LandingPageModule { }
