+++ src/app/mainModules/services/services.component.html
@@ -1,0 +1,55 @@
+<div class="container" [@slideFromTop]>
+
+    <h1>{{ 'OUR_SERVICES' | translate }}</h1>
+    <div class="row">
+        <div class="service">
+            <i class="fa-solid fa-location-dot"></i>
+            <h2>{{ 'SPECIFIC_ADDRESSES' | translate }}</h2>
+            <p>{{ 'ADDRESSES_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-sharp fa-solid fa-chart-simple"></i>
+            <h2>{{ 'MARKETING' | translate }}</h2>
+            <p>{{ 'MARKETING_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-solid fa-magnifying-glass-chart"></i>
+            <h2>{{ 'DATA_ANALYSIS' | translate }}</h2>
+            <p>{{ 'DATA_ANALYSIS_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-sharp fa-solid fa-universal-access"></i>
+            <h2>{{ 'WCAG' | translate }}</h2>
+            <p>{{ 'WCAG_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-solid fa-code"></i>
+            <h2>{{ 'WEB_DESIGN' | translate }}</h2>
+            <p>{{ 'WEB_DESIGN_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-solid fa-file-lines"></i>
+            <h2>{{ 'BLOG' | translate }}</h2>
+            <p>{{ 'BLOG_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-solid fa-house"></i>
+            <h2>{{ 'OPPORTUNITY_TRACKING' | translate }}</h2>
+            <p>{{ 'OPPORTUNITY_TRACKING_DESCRIPTION' | translate }}</p>
+        </div>
+
+        <div class="service">
+            <i class="fa-solid fa-mobile-screen-button"></i>
+            <h2>{{ 'COMPUTER_MOBILE' | translate }}</h2>
+            <p>{{ 'COMPUTER_MOBILE_DESCRIPTION' | translate }}</p>
+        </div>
+    </div>
+
+
+</div>
