+++ src/app/mainModules/services/services.module.ts
@@ -22,6 +22,7 @@ export function HttpLoaderFactory(http: HttpClient) {
         deps: [HttpClient]
       }
     })
-  ]
+  ],
+  exports:[ServicesComponent]
 })
 export class ServicesModule { }
