--- src/app/services/services.component.html
@@ -1,55 +1,0 @@
-<div class="container" [@slideFromTop]>
-
-    <h1>Our services</h1>
-    <div class="row">
-        <div class="service">
-            <i class="fa-solid fa-location-dot"></i>
-            <h2>Specific addresses</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-sharp fa-solid fa-chart-simple"></i>
-            <h2>Marketing</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-solid fa-magnifying-glass-chart"></i>
-            <h2>Data Analysis</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-sharp fa-solid fa-universal-access"></i>
-            <h2>WCAG</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-solid fa-code"></i>
-            <h2>Web Design</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-solid fa-file-lines"></i>
-            <h2>Newsletter/blog</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-solid fa-house"></i>
-            <h2>Opportunity tracking</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-
-        <div class="service">
-            <i class="fa-solid fa-mobile-screen-button"></i>
-            <h2>Laptop-Mobile</h2>
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing.</p>
-        </div>
-    </div>
-
-
-</div>
