+++ src/app/shared/components/subscribe-popup/subscribe-popup.component.scss
@@ -8,7 +8,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    z-index: 9999;
+    z-index: 3;
   }
   
   .popup-container {
