+++ src/app/shared/shared.module.ts
@@ -1,0 +1,53 @@
+import { NgModule } from '@angular/core';
+import {TranslateModule, TranslateLoader} from '@ngx-translate/core';
+import {TranslateHttpLoader} from '@ngx-translate/http-loader';
+import {HttpClientModule, HttpClient} from '@angular/common/http';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { FooterComponent } from './components/footer/footer.component';
+import { AppLoaderComponent } from './components/app-loader/app-loader.component';
+import { HeaderComponent } from './components/header/header.component';
+import { SubscribePopupComponent } from './components/subscribe-popup/subscribe-popup.component';
+import { CommonModule } from '@angular/common';
+import { LoginComponent } from './components/login/login.component';
+import { SigninComponent } from './components/signin/signin.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { AppRoutingModule } from './app-routing.module';
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
+
+@NgModule({
+  declarations: [
+    HeaderComponent,
+    FooterComponent,
+    AppLoaderComponent,
+    SubscribePopupComponent,
+    LoginComponent,
+    SigninComponent
+  ],
+  imports: [
+    ReactiveFormsModule,
+    AppRoutingModule,
+    CommonModule,
+    HttpClientModule,
+    BrowserAnimationsModule,
+    TranslateModule.forRoot({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      }
+    })
+  ],
+  exports: [
+    HeaderComponent,
+    FooterComponent,
+    AppLoaderComponent,
+    SubscribePopupComponent,
+    LoginComponent,
+    SigninComponent,
+    AppRoutingModule
+  ]
+})
+export class SharedModule { }
